+++ store/tikv/tests/util_test.go
@@ -18,7 +18,6 @@ import (
 	"flag"
 	"fmt"
 	"strings"
-	"sync"
 	"unsafe"
 
 	. "github.com/pingcap/check"
@@ -28,14 +27,13 @@ import (
 	"github.com/pingcap/tidb/store/mockstore/unistore"
 	"github.com/pingcap/tidb/store/tikv"
 	"github.com/pingcap/tidb/store/tikv/config"
+	"github.com/pingcap/tidb/store/tikv/util"
 	"github.com/pingcap/tidb/store/tikv/util/codec"
 	pd "github.com/tikv/pd/client"
 )
 
 var (
-	withTiKVGlobalLock sync.RWMutex
-	WithTiKV           = flag.Bool("with-tikv", false, "run tests with TiKV cluster started. (not use the mock server)")
-	pdAddrs            = flag.String("pd-addrs", "127.0.0.1:2379", "pd addrs")
+	pdAddrs = flag.String("pd-addrs", "127.0.0.1:2379", "pd addrs")
 )
 
 // NewTestStore creates a KVStore for testing purpose.
@@ -44,7 +42,7 @@ func NewTestStore(c *C) *tikv.KVStore {
 		flag.Parse()
 	}
 
-	if *WithTiKV {
+	if *util.WithTiKV {
 		addrs := strings.Split(*pdAddrs, ",")
 		pdClient, err := pd.NewClient(addrs, pd.SecurityOption{})
 		c.Assert(err, IsNil)
@@ -86,23 +84,7 @@ func clearStorage(store *tikv.KVStore) error {
 }
 
 // OneByOneSuite is a suite, When with-tikv flag is true, there is only one storage, so the test suite have to run one by one.
-type OneByOneSuite struct{}
-
-func (s *OneByOneSuite) SetUpSuite(c *C) {
-	if *WithTiKV {
-		withTiKVGlobalLock.Lock()
-	} else {
-		withTiKVGlobalLock.RLock()
-	}
-}
-
-func (s *OneByOneSuite) TearDownSuite(c *C) {
-	if *WithTiKV {
-		withTiKVGlobalLock.Unlock()
-	} else {
-		withTiKVGlobalLock.RUnlock()
-	}
-}
+type OneByOneSuite = util.OneByOneSuite
 
 func encodeKey(prefix, s string) []byte {
 	return codec.EncodeBytes(nil, []byte(fmt.Sprintf("%s_%s", prefix, s)))
