+++ store/tikv/tikv_test.go
@@ -14,36 +14,11 @@
 package tikv
 
 import (
-	"flag"
-	"sync"
-
 	. "github.com/pingcap/check"
+	"github.com/pingcap/tidb/store/tikv/util"
 )
 
-var (
-	withTiKVGlobalLock sync.RWMutex
-	WithTiKV           = flag.Bool("with-tikv", false, "run tests with TiKV cluster started. (not use the mock server)")
-)
-
-// OneByOneSuite is a suite, When with-tikv flag is true, there is only one storage, so the test suite have to run one by one.
-type OneByOneSuite struct{}
-
-func (s *OneByOneSuite) SetUpSuite(c *C) {
-	if *WithTiKV {
-		withTiKVGlobalLock.Lock()
-	} else {
-		withTiKVGlobalLock.RLock()
-	}
-}
-
-func (s *OneByOneSuite) TearDownSuite(c *C) {
-	if *WithTiKV {
-		withTiKVGlobalLock.Unlock()
-	} else {
-		withTiKVGlobalLock.RUnlock()
-	}
-}
-
+type OneByOneSuite = util.OneByOneSuite
 type testTiKVSuite struct {
 	OneByOneSuite
 }
